Fixed 404 download link on index html page css changes
I changed the width of the page to max width 600px to accommodate a PRE
style that better represents “code” commands and max width to better
accommodate mobile devices. I made many more changes to mobile stylings
for the page so it scales really well for smaller devices now. I also
proof-read the content of the page and corrected any English translation
errors for you and unified some of the messages you had on the page
adding links where appropriate.